+++ kubernetes-model/kubernetes-model/src/main/java/io/fabric8/kubernetes/api/model/Quantity.java
@@ -31,6 +31,7 @@ import lombok.ToString;
 
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.math.MathContext;
 import java.util.HashMap;
 import java.util.Map;
 import java.io.Serializable;
@@ -98,7 +99,7 @@ public class Quantity  implements Serializable {
     this.format = format;
   }
 
-  public static BigDecimal getAmountInBytes(Quantity quantity) {
+  public static BigDecimal getAmountInBytes(Quantity quantity) throws ArithmeticException {
     String value = "";
     if (quantity.getAmount() != null && quantity.getFormat() != null) {
       value = quantity.getAmount() + quantity.getFormat();
@@ -130,49 +131,49 @@ public class Quantity  implements Serializable {
 
     switch (formatStr) {
       case "Ki":
-        multiple = binaryFactor.pow(10);
+        multiple = binaryFactor.pow(10, MathContext.DECIMAL64);
         break;
       case "Mi":
-        multiple = binaryFactor.pow(20);
+        multiple = binaryFactor.pow(20, MathContext.DECIMAL64);
         break;
       case "Gi":
-        multiple = binaryFactor.pow(30);
+        multiple = binaryFactor.pow(30, MathContext.DECIMAL64);
         break;
       case "Ti":
-        multiple = binaryFactor.pow(40);
+        multiple = binaryFactor.pow(40, MathContext.DECIMAL64);
         break;
       case "Pi":
-        multiple = binaryFactor.pow(50);
+        multiple = binaryFactor.pow(50, MathContext.DECIMAL64);
         break;
       case "Ei":
-        multiple = binaryFactor.pow(60);
+        multiple = binaryFactor.pow(60, MathContext.DECIMAL64);
         break;
       case "n":
-        multiple = decimalFactor.pow(-9);
+        multiple = decimalFactor.pow(-9, MathContext.DECIMAL64);
         break;
       case "u":
-        multiple = decimalFactor.pow(-6);
+        multiple = decimalFactor.pow(-6, MathContext.DECIMAL64);
         break;
       case "m":
-        multiple = decimalFactor.pow(-3);
+        multiple = decimalFactor.pow(-3, MathContext.DECIMAL64);
         break;
       case "k":
-        multiple = decimalFactor.pow(3);
+        multiple = decimalFactor.pow(3, MathContext.DECIMAL64);
         break;
       case "M":
-        multiple = decimalFactor.pow(6);
+        multiple = decimalFactor.pow(6, MathContext.DECIMAL64);
         break;
       case "G":
-        multiple = decimalFactor.pow(9);
+        multiple = decimalFactor.pow(9, MathContext.DECIMAL64);
         break;
       case "T":
-        multiple = decimalFactor.pow(12);
+        multiple = decimalFactor.pow(12, MathContext.DECIMAL64);
         break;
       case "P":
-        multiple = decimalFactor.pow(15);
+        multiple = decimalFactor.pow(15, MathContext.DECIMAL64);
         break;
       case "E":
-        multiple = decimalFactor.pow(18);
+        multiple = decimalFactor.pow(18, MathContext.DECIMAL64);
         break;
     }