+++ openshift-client/src/main/java/io/fabric8/openshift/client/DefaultOpenShiftClient.java
@@ -456,6 +456,9 @@ public class DefaultOpenShiftClient extends BaseClient implements NamespacedOpen
   public BatchAPIGroupDSL batch() { return adapt(BatchAPIGroupClient.class); }
 
   @Override
+  public MetricAPIGroupDSL top() { return adapt(MetricAPIGroupClient.class); }
+
+  @Override
   public PolicyAPIGroupDSL policy() { return adapt(PolicyAPIGroupClient.class); }
 
   @Override